refactor: moved styled cells
