+++ src/components/layout/StyledTableCell.tsx
