+++ src/components/task/TaskRow.tsx
@@ -2,8 +2,8 @@ import React, { ChangeEvent, KeyboardEvent } from 'react';
 import './TaskRow.css';
 import { Task, Time } from '../../app/redux';
 import { TableRow, Input, Button } from '@material-ui/core';
-import IconTableCell from '../IconTableCell';
-import StyledTableCell from '../StyledTableCell';
+import IconTableCell from '../layout/IconTableCell';
+import StyledTableCell from '../layout/StyledTableCell';
 import HighlightOffIcon from '@material-ui/icons/HighlightOff';
 import TimeCellRow from '../time/TimeCells';
 import SummaryCell from '../time/SummaryCell';
